% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_gradient_ll}
\alias{prox_gradient_ll}
\title{Proxmal-gradient log-likelihood estimator.}
\usage{
prox_gradient_ll(data, theta_i, regularizer)
}
\arguments{
\item{data}{N x P matrix corresponding to the raw data.}

\item{theta_i}{Estimate for precision.}

\item{regularizer}{Regularizing constant, lambda.}
}
\value{
Log-likelihood estimate.
}
\description{
Estimates the log-likeihood for the corresponding
             precision matrix and data set.
}
\author{
\packageMaintainer{changepointsHD}
}
