\name{changepointsHD-package}
\alias{changepointsHD-package}
\alias{changepointsHD}
\docType{package}
\title{
\packageTitle{changepointsHD}
}
\description{
\packageDescription{changepointsHD}
}
\details{
\tabular{ll}{
Package: \tab changepointsHD\cr
Type: \tab Package\cr
Version: \tab 0.3.0\cr
Date: \tab 2017-11-06\cr
License: \tab GPL-2\cr
}
}
%\details{
%\tabular{ll}{
%Package: \tab \packagePackage{changepoints} \cr
%Type: \tab \packageType{changepoints} \cr
%Version: \tab \packageVersion{changepoints} \cr
%Date: \tab \packageDate{changepoints} \cr
%Licence: \tab packageLicense{changepoints} \cr
%}
\author{
\packageAuthor{changepointsHD}

Maintainer: \packageMaintainer{changepointsHD}
}
\references{1. L. Bybee and Y. Atchade: "A Scalable Algorithm for Gaussian Graphical Models with Change-Points", 2017; [http://arxiv.org/abs/1707.04306 arXiv:1707.04306].
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
%~~ in the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
